Add files via upload
Initial commit to a crude Google scraper for a research project I am currently working on.

First it searches Google for articles by keywords and extracts their URLs. Then the script goes through each URL looking for comments and stores them in a file. It does not work with dynamic comment sections (for example Disqus) and those requiring authorization.